// frontend/src/api/model/getFilesOwnersKindFilesOrderBy.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export type GetFilesOwnersKindFilesOrderBy = typeof GetFilesOwnersKindFilesOrderBy[keyof typeof GetFilesOwnersKindFilesOrderBy];


export const GetFilesOwnersKindFilesOrderBy = {
  created_at: 'created_at',
  size: 'size',
  name: 'name',
} as const;