// frontend/src/api/model/githubCom4H1RZooraInternalDomainOwnerFile.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Zoora API
 * REST API for the Zoora education platform.
 * OpenAPI spec version: 1.0
 */

export interface GithubCom4H1RZooraInternalDomainOwnerFile {
  created_at?: string;
  deletable?: boolean;
  id?: string;
  mime_type?: string;
  /** media model_type, or "recording" */
  model_type?: string;
  name?: string;
  size?: number;
  /** "media" | "recording" */
  source?: string;
}